ai-image-util/pyproject.toml
[project]
name = "ai-image-util"
version = "0.1.0"
description = "AI Image Utility for generating 19th-century engraving prompts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow<12.2",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "tqdm>=4.67.3",
]

[project.scripts]
ai-image-gen = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "hatchling>=1.29.0",
]
